# shifter/shifter_platform/openapi/v1.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Shifter Platform API",
    "version": "v1",
    "description": "Authenticated HTTP/JSON API for Shifter platform integrations."
  },
  "paths": {
...
    "/api/v1/mission-control/credentials/": {
      "post": {
        "operationId": "api_v1_mission_control_credentials_create",
        "description": "Validate and persist a Mission Control credential.",
        "tags": [
          "mission-control"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CredentialCreate"
              }
            },
            "application/x-www-form-urlencoded": {
              "schema": {
                "$ref": "#/components/schemas/CredentialCreate"
              }
            },
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/CredentialCreate"
              }
            }
          },
          "required": true
        },
        "security": [
          {
            "ApiTokenAuth": []
          },
          {
            "cookieAuth": []
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CredentialCreateResponse"
                }
              }
            },
            "description": ""
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            },
            "description": "Invalid request."
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            },
            "description": "Authentication failed."
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            },
            "description": "Permission denied."
          }
        },
        "x-required-scopes": [
          "mission_control:credentials:write"
        ]
      }
    },
    "/api/v1/mission-control/credentials/{credential_id}/delete/": {
      "post": {
        "operationId": "api_v1_mission_control_credentials_delete",
        "description": "Delete a credential visible to the authenticated actor.",
        "parameters": [
          {
            "in": "path",
            "name": "credential_id",
            "schema": {
              "type": "integer"
            },
            "required": true
          }
        ],
        "tags": [
          "mission-control"
        ],
        "security": [
          {
            "ApiTokenAuth": []
          },
          {
            "cookieAuth": []
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SuccessResponse"
                }
              }
            },
            "description": ""
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            },
            "description": "Invalid request."
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            },
            "description": "Authentication failed."
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            },
            "description": "Permission denied."
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            },
            "description": "Resource not found."
          }
        },
        "x-required-scopes": [
          "mission_control:credentials:write"
        ]
      }
    },
...
    "/api/v1/mission-control/ngfw/": {
      "post": {
        "operationId": "api_v1_mission_control_ngfw_create",
        "description": "Start NGFW provisioning for the authenticated actor.",
        "tags": [
          "mission-control"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/NGFWCreate"
              }
            },
            "application/x-www-form-urlencoded": {
              "schema": {
                "$ref": "#/components/schemas/NGFWCreate"
              }
            },
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/NGFWCreate"
              }
            }
          }
        },
        "security": [
          {
            "ApiTokenAuth": []
          },
          {
            "cookieAuth": []
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/NGFWCreateResponse"
                }
              }
            },
            "description": ""
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            },
            "description": "Invalid request."
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            },
            "description": "Authentication failed."
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            },
            "description": "Permission denied."
          },
          "429": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            },
            "description": "Request was throttled."
          }
        },
        "x-required-scopes": [
          "mission_control:ngfw:write"
        ]
      }
    },
    "/api/v1/mission-control/ngfw/{app_id}/destroy/": {
      "post": {
        "operationId": "api_v1_mission_control_ngfw_destroy",
        "description": "Start NGFW deprovisioning for the requested app id.",
        "parameters": [
          {
            "in": "path",
            "name": "app_id",
            "schema": {
              "type": "string",
              "format": "uuid"
            },
            "required": true
          }
        ],
        "tags": [
          "mission-control"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/NGFWDestroy"
              }
            },
            "application/x-www-form-urlencoded": {
              "schema": {
                "$ref": "#/components/schemas/NGFWDestroy"
              }
            },
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/NGFWDestroy"
              }
            }
          }
        },
        "security": [
          {
            "ApiTokenAuth": []
          },
          {
            "cookieAuth": []
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/NGFWDestroyResponse"
                }
              }
            },
            "description": ""
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            },
            "description": "Invalid request."
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            },
            "description": "Authentication failed."
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            },
            "description": "Permission denied."
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            },
            "description": "Resource not found."
          }
        },
        "x-required-scopes": [
          "mission_control:ngfw:write"
        ]
      }
    },
...
    "/api/v1/mission-control/ngfw/list/": {
      "get": {
        "operationId": "api_v1_mission_control_ngfw_list",
        "description": "Return NGFWs owned by the authenticated actor.",
        "tags": [
          "mission-control"
        ],
        "security": [
          {
            "ApiTokenAuth": []
          },
          {
            "cookieAuth": []
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/NGFWListResponse"
                }
              }
            },
            "description": ""
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            },
            "description": "Invalid request."
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            },
            "description": "Authentication failed."
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            },
            "description": "Permission denied."
          }
        },
        "x-required-scopes": [
          "mission_control:ngfw:read"
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ApiError": {
        "type": "object",
        "description": "Schema-only component for the canonical platform API error envelope.\n\nMirrors :func:`shared.api.errors.api_exception_handler` /\n:func:`shared.api.errors.api_error_response` output. It describes the wire\ncontract; it never runs at request time.",
        "properties": {
          "error": {
            "$ref": "#/components/schemas/ApiErrorBody"
          }
        },
        "required": [
          "error"
        ]
      },
      "ApiErrorBody": {
        "type": "object",
        "description": "Schema-only description of the ``shared.api.errors`` envelope body.",
        "properties": {
          "code": {
            "type": "string",
            "description": "Stable machine-readable error code."
          },
          "message": {
            "type": "string",
            "description": "Safe, user-facing error message."
          },
          "details": {
            "description": "Optional structured field-level detail (present on validation errors)."
          },
          "request_id": {
            "type": "string",
            "description": "Correlation id echoed from the X-Request-ID request header, when present."
          }
        },
        "required": [
          "code",
          "message"
        ]
      },
...
      "CredentialCreate": {
        "type": "object",
        "description": "Validate credential creation requests before schema-specific validation.",
        "properties": {
          "credential_type": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "expires_at": {
            "type": "string",
            "nullable": true
          },
          "scm_folder_name": {
            "type": "string"
          },
          "scm_pin_id": {
            "type": "string"
          },
          "scm_pin_value": {
            "type": "string"
          },
          "sls_region": {
            "type": "string"
          },
          "authcode": {
            "type": "string"
          }
        },
        "required": [
          "credential_type"
        ]
      },
      "CredentialCreateResponse": {
        "type": "object",
        "description": "Response body for ``CredentialCreateView.post`` (HTTP 201).",
        "properties": {
          "id": {
            "type": "integer"
          },
          "name": {
            "type": "string"
          },
          "credential_type": {
            "type": "string"
          }
        },
        "required": [
          "credential_type",
          "id",
          "name"
        ]
      },
...
      "NGFWCreate": {
        "type": "object",
        "description": "Validate NGFW creation requests.",
        "properties": {
          "name": {
            "type": "string",
            "default": ""
          },
          "deployment_profile_id": {
            "type": "integer",
            "nullable": true
          },
          "registration_method": {
            "type": "string",
            "default": ""
          },
          "scm_credential_id": {
            "type": "integer",
            "nullable": true
          },
          "otp_value": {
            "type": "string",
            "nullable": true
          },
          "otp_folder": {
            "type": "string",
            "nullable": true
          }
        }
      },
      "NGFWCreateResponse": {
        "type": "object",
        "description": "Response body for ``NGFWCreateView.post`` (HTTP 201).",
        "properties": {
          "id": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "status": {
            "type": "string"
          }
        },
        "required": [
          "id",
          "name",
          "status"
        ]
      },
      "NGFWDestroy": {
        "type": "object",
        "description": "Validate NGFW destroy requests.",
        "properties": {
          "confirm_name": {
            "type": "string",
            "default": ""
          }
        }
      },
      "NGFWDestroyResponse": {
        "type": "object",
        "description": "Response body for ``NGFWDestroyView.post``.",
        "properties": {
          "status": {
            "type": "string"
          }
        },
        "required": [
          "status"
        ]
      },
      "NGFWListItem": {
        "type": "object",
        "description": "One entry from ``NGFWListView.get``.",
        "properties": {
          "id": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "status": {
            "type": "string"
          },
          "created_at": {
            "type": "string",
            "format": "date-time"
          },
          "serial_number": {
            "type": "string",
            "nullable": true
          }
        },
        "required": [
          "created_at",
          "id",
          "name",
          "serial_number",
          "status"
        ]
      },
      "NGFWListResponse": {
        "type": "object",
        "description": "Response body for ``NGFWListView.get``.",
        "properties": {
          "ngfws": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/NGFWListItem"
            }
          }
        },
        "required": [
          "ngfws"
        ]
      },
...
      "SuccessResponse": {
        "type": "object",
        "description": "Generic ``{\"success\": true}`` response body.\n\nShared by the range lifecycle mutations (cancel/destroy/pause/resume),\nupload cancel, and credential delete — every Mission Control mutation\nendpoint that has no payload beyond a boolean acknowledgement.",
        "properties": {
          "success": {
            "type": "boolean"
          }
        },
        "required": [
          "success"
        ]
      },
...
    },
    "securitySchemes": {
      "ApiTokenAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "shf",
        "description": "Platform API token with scopes from shared.api_tokens.scopes."
      },
      "cookieAuth": {
        "type": "apiKey",
        "in": "cookie",
        "name": "sessionid"
      }
    }
  }
}